Make API more DRY
